+++ src/main/java/com/nc/unc/repositories/impl/OrderItemRepository.java
@@ -2,7 +2,7 @@ package com.nc.unc.repositories.impl;
 
 import com.nc.unc.model.OrderItem;
 
-public class OrderItemRepository extends RepositoryEntity<Long, OrderItem> {
+public class OrderItemRepository extends RepositoryEntity<Integer, OrderItem> {
 
     public OrderItemRepository() {
         super(OrderItemRepository.class);
